pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-out"
authors = [
    {name = "RenoirTan"}
]
description = "Manage your local git repository mirrors."
readme = "README.md"
requires-python = ">=3.7"
keywords = ["git", "clone", "mirror", "bare", "backup", "repository", "repositories"]
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
	"Intended Audience :: Developers",
	"Topic :: Software Development :: Version Control",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: System :: Archiving :: Mirroring",
	"Topic :: Utilities"
]
dependencies = [
    "PyGithub>=1.58.1",
    "GitPython>=3.1.31"
]
dynamic = ["version"]

[project.scripts]
gitout-mirror = "gitout.mirror:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gitout"]

[tool.setuptools.dynamic.version]
attr = "gitout._data.VERSION"